+++ zipapps/__main__.py
@@ -205,23 +205,28 @@ def main():
         '--layer-mode',
         action='store_true',
         dest='layer_mode',
-        help=
-        'Layer mode for the serverless use case, __main__.py / ensure_zipapps.py / activate_zipapps.py files will not be set in this mode.'
+        help='Layer mode for the serverless use case, '
+        '__main__.py / ensure_zipapps.py / activate_zipapps.py files will not be set in this mode.'
     )
-    parser.add_argument(
-        '--layer-mode-prefix',
-        default='python',
-        dest='layer_mode_prefix',
-        help=
-        'Only work while --layer-mode is set, will move the files in the given prefix folder.'
-    )
+    parser.add_argument('--layer-mode-prefix',
+                        default='python',
+                        dest='layer_mode_prefix',
+                        help='Only work while --layer-mode is set, '
+                        'will move the files in the given prefix folder.')
     parser.add_argument(
         '-czc',
         '--clear-zipapps-cache',
         action='store_true',
         dest='clear_zipapps_cache',
-        help=
-        'Clear the zipapps cache folder after running, but maybe failed for .pyd/.so files.',
+        help='Clear the zipapps cache folder after running, '
+        'but maybe failed for .pyd/.so files.',
+    )
+    parser.add_argument(
+        '--chmod',
+        default='',
+        dest='chmod',
+        help='os.chmod(int(chmod, 8)) for unzip files with `--chmod=777`,'
+        ' unix-like system only',
     )
     if len(sys.argv) == 1:
         return parser.print_help()
@@ -255,6 +260,7 @@ def main():
         layer_mode_prefix=args.layer_mode_prefix,
         clear_zipapps_cache=args.clear_zipapps_cache,
         unzip_exclude=args.unzip_exclude,
+        chmod=args.chmod,
     )